+++ components/Cart/index.tsx
@@ -22,36 +22,6 @@ const Cart = () => {
   return (
     <>
       <h1 className="mb-3 text-xl">Our Cart</h1>
-      <div className="mb-4 flex w-1/3">
-        <button
-          onClick={() => dispatch(clearCart())}
-          className="w-1/2 rounded-full bg-blue-500 py-2 px-4 font-bold text-white hover:bg-blue-700"
-        >
-          Clear Cart
-        </button>
-        <button
-          onClick={() =>
-            dispatch(
-              addItem({
-                id: 3,
-                title: 'Face Wash',
-                img: 'https://picsum.photos/id/237/200/300',
-                count: 1,
-                price: 10,
-              }),
-            )
-          }
-          className="w-1/2 rounded-full bg-blue-500 py-2 px-4 font-bold text-white hover:bg-blue-700"
-        >
-          Add Face Wash
-        </button>
-        <button
-          onClick={() => dispatch(removeItem({ id: 3, title: 'Face wash' }))}
-          className="ml-1 w-1/2 rounded-full bg-blue-500 py-2 px-4 font-bold text-white hover:bg-blue-700"
-        >
-          Remove Face Wash
-        </button>
-      </div>
 
       <h1>Total: {total}</h1>
       <ul className="flex mb-4 flex-col pl-0">
